Extract useActionFromRoute hook and clean up duplicated action lookup logic
