+++ clients/privacy-center/components/privacy-request/VerificationPage.tsx
@@ -4,10 +4,8 @@ import { Flex } from "fidesui";
 import { useParams, useRouter } from "next/navigation";
 import React, { useEffect, useState } from "react";
 
-import { decodePolicyKey } from "~/common/policy-key";
+import useActionFromRoute from "~/common/hooks/useActionFromRoute";
 import { ModalViews, VerificationType } from "~/components/modals/types";
-import { useConfig } from "~/features/common/config.slice";
-import { PrivacyRequestOption } from "~/types/config";
 
 import VerificationForm from "../modals/verification-request/VerificationForm";
 
@@ -16,26 +14,12 @@ type VerificationPageProps = {
 };
 
 const VerificationPage = ({ actionKey }: VerificationPageProps) => {
-  const config = useConfig();
   const router = useRouter();
   const params = useParams();
   const propertyPath = params?.propertyPath as string | undefined;
   const basePath = propertyPath ? `/${propertyPath}` : "";
   const [privacyRequestId, setPrivacyRequestId] = useState<string>("");
-
-  const decoded = decodePolicyKey(actionKey);
-  const colonIndex = decoded.indexOf(":");
-  const actionIndex =
-    colonIndex !== -1 ? parseInt(decoded.slice(0, colonIndex), 10) : NaN;
-  const policyKey =
-    colonIndex !== -1 ? decoded.slice(colonIndex + 1) : decoded;
-  const actions = config.actions ?? [];
-  const action = (
-    !Number.isNaN(actionIndex) &&
-    actions[actionIndex]?.policy_key === policyKey
-      ? actions[actionIndex]
-      : actions.find((a) => a.policy_key === policyKey)
-  ) as PrivacyRequestOption | undefined;
+  const action = useActionFromRoute(actionKey);
 
   useEffect(() => {
     if (typeof window !== "undefined") {
